move prompt model into db types
